requirements.txt
crewai>=0.11.0
fastapi>=0.104.1
uvicorn>=0.24.0
streamlit>=1.29.0
python-dotenv>=1.0.0
pandas>=2.1.3
yfinance>=0.2.33
alpha_vantage>=2.3.1
google-generativeai>=0.3.2
litellm>=1.30.7 